+++ README.md
@@ -1,6 +1,7 @@
-# Lua IPython Kernel
+# iPyLua: Lua IPython Kernel
 
-This is a kernel to support Lua with [IPython](http://ipython.org).  It is pure Lua and should work with both Lua and LuaJIT.
+This is a kernel to support Lua with [IPython](http://ipython.org).  It is pure
+Lua and should work with both Lua and LuaJIT.
 
 
 ## Requirements
@@ -9,51 +10,35 @@ The following Lua libraries are required:
 
  * [lzmq](https://github.com/zeromq/lzmq) zeromq for Lua
 
-The code for Lua JSON of dkjs habeen embedded into the
-Kenel source. Uuid's are generated using SO commands via
-io.popen.
-
-Here's how to install via [LuaRocks](http://luarocks.org/):
-    
-```
-# You need zeromq... on OSX I use Homebrew:
-# brew install zeromq
-# On Ubuntu:
-# sudo apt-get install libzmq-dev
-
-# The LuaRocks
-sudo luarocks install https://raw.github.com/Neopallium/lua-zmq/master/rockspecs/lua-zmq-scm-1.rockspec
-sudo luarocks install dkjson
-sudo luarocks install uuid
-```
-
-Of course you also need to [install IPython](http://ipython.org/install.html)... 
-
+The code for Lua JSON of [dkjson](http://dkolf.de/src/dkjson-lua.fsl/home) and
+[uuid](https://github.com/Tieske/uuid/blob/master/src/uuid.lua) has been copied
+into this module to avoid external dependencies.
 
 ## Installation
 
 The installation process is janky right now.
 
- * Install the Requirements above
+ * Install the Requirements: [zeromq](http://zeromq.org/) and
+   [lzmq](https://github.com/zeromq/lzmq).
 
- * Create a profile with IPython
+ * Create a profile with IPython:
 
 ```
-ipython profile create lua
+$ ipython profile create lua
 ```
 
- * Modify the profile's `ipython_config.py` to use lua_ipython_kernel.  This
+ * Modify the profile's `ipython_config.py` to use iPyLua.  This
  will be at either `~/.config/ipython/profile_lua/ipython_config.py` or
  `~/.ipython/profile_lua/ipython_config.py`:
 
-```
+```Python
 # Configuration file for ipython.
    
 c = get_config()
    
 c.KernelManager.kernel_cmd = [
-    "luajit",              # select your Lua interpreter here
-    "ipython_kernel.lua",  # probably need full path
+    "luajit", # select your Lua interpreter here (lua5.2, lua5.1, luajit)
+    "iPyLua/iPyLuaKernel.lua",  # probably need full path
     "{connection_file}"
 ]
    
@@ -65,9 +50,9 @@ c.Session.keyfile = b''
  * Invoke IPython with this Lua kernel:
 
 ```
-ipython console --profile lua
+$ ipython console --profile lua
 # or 
-ipython notebook --profile lua
+$ ipython notebook --profile lua
 ```
 
 ## TODO
@@ -81,23 +66,43 @@ ipython notebook --profile lua
 
 ## Acknowledgements
 
-Thanks to Andrew Gibiansky for his [IHaskell article](http://andrew.gibiansky.com/blog/ipython/ipython-kernels/) that inspired this.  
+Thanks to Evan Wies who has written the original code
+[lua_ipython_kernel](https://github.com/neomantra/lua_ipython_kernel).
 
-Thanks to the makers of the dependencies of this library, who made this pretty easy to create: [Robert Jakabosky](https://github.com/Neopallium), [David Kolf](http://dkolf.de/src/dkjson-lua.fsl/home), and [Thijs Schreijer](https://github.com/Tieske).  
+Thanks to Andrew Gibiansky for his
+[IHaskell article](http://andrew.gibiansky.com/blog/ipython/ipython-kernels/)
+that inspired this.
+
+Thanks to the makers of the dependencies of this library, who made this pretty
+easy to create: [Robert Jakabosky](https://github.com/Neopallium),
+[David Kolf](http://dkolf.de/src/dkjson-lua.fsl/home), and
+[Thijs Schreijer](https://github.com/Tieske).
 
 And of course thanks to the [IPython folks ](http://ipython.org/citing.html).
 
-
 ## LICENSE
 
-**lua_ipython_kernel** is distributed under the [MIT License](http://opensource.org/licenses/mit-license.php).
+**lua_ipython_kernel** is distributed under the
+  [MIT License](http://opensource.org/licenses/mit-license.php).
 
 > lua_ipython_kernel
 > 
 > Copyright (c) 2013 Evan Wies.  All rights reserved.
 > 
-> Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+> Permission is hereby granted, free of charge, to any person obtaining a copy
+> of this software and associated documentation files (the "Software"), to deal
+> in the Software without restriction, including without limitation the rights
+> to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+> copies of the Software, and to permit persons to whom the Software is
+> furnished to do so, subject to the following conditions:
 > 
-> The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+> The above copyright notice and this permission notice shall be included in all
+> copies or substantial portions of the Software.
 > 
-> THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+> THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+> IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+> FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+> AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+> LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+> OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+> SOFTWARE.
